avoid Voip::get_subscriber functions where possible
This needs to be cleaned up far more than expected. The code uses
subscriber information out of $c->session->{user} and sometimes
$c->session->{user}{data} at random as it seems. It's not clear what's
the difference between the logged-in subscriber and the one we operate
on, even though currently they are always identical. Further we should
use the user-provided credentials to access the provisioning interface
and remove all the custom access functions in the Provisioning Model.